Encapsulate DiviK into package
