+++ divik/cluster/_divik/__init__.py
@@ -1,0 +1,3 @@
+"""DiviK algorithm implementation."""
+
+from ._sklearn import DiviK
