+++ test/cluster/divik/test_divik.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 import numpy as np
 import numpy.testing as npt
 
-import divik.cluster._divik as dv
+import divik.cluster._divik._core as dv
 
 
 def returns(value):
